New: remove methods that were moved to the `view_utils` module
The methods are: `_make_fields_dict()`, `_make_related_dict()`,
`make_detail_dict()`. The methods in the `view_utils` module are more clear, and
easier to maintain, and can be reused for other views.

These methods will be used in the GenericModelViewSet class.

Also removed `get_model_name()` method, was not necessary anymore.

Refs #127