New: add contribution_info_mixin.py, contains functions to handle Contributions
The mixin can is added to MusicalWork/Section/Part to handle the relationship to
Contributions. It allows easy access to sets of Persons, GeographicAreas, and
dates based on the role of the Contribution. There is also a function to
determine the overall certainty of a MusicalWork/Section/Part based on the
certainty of all Contributions to that MusicalWork/Section/Part.